PSD dependency list now changes if using androidx
When adding libraries using the PSD, there is a list that we show to the
user, this list contains all libraries from GoogleMavenArtifactId that
are available in the maven repository, including androidx and old (non
androidx) support libraries.

This CL filters out androidx libraries when only old versions are in
use or filters old versions when at least one of the androidx support
libraries is in use.

Bug: 76454148
Test: Manually checked that the list are correct in both cases (when
androidx is used or not).

Change-Id: I30671fe848e154b58e060db4d0499ee16394ad15