Disable auto-profiling when manually stopping/switching sessions
This CL introduces a new boolean in StudioProfilers to control
whether the profilers should auto select the suitable device+process
combination to start profiling. If set to false, (as in the case
when the user explicitly stops or selects a seession), the profilers
will not try to start profiling even when it detects new devices and
processes.

Bug: 77658332
Test: updated
Change-Id: Ie09c6c627c41daa4731b915bbf0c48ac99b9b214