waitForRenderToFinish after dragging a component
NlEditorTest.basicLayoutEdit is flaky because sometimes the Button tag
was not showing in xml file after dragging a component. Addressing this
issue by waiting for render to finish after dragging a component.

Another thing observed: TextView is already in the sample project, it is
not valid to verify TextView in the xml file, instead, use
Layouts/TableRow in test.

Bug: 77856035
Test: Run NlEditorTest.basicLayoutEdit with gradle locally
Change-Id: I63ee53fc2d2a25489026cf1fce5e01dc5841e7c6