Use FakeAdbServer for createAndRunInstantApp
This commit is the first step in bringing sanity tests to the next level
of stability by faking the emulator. The emulator's startup time can
vary greatly across builds and times of the day. This variance made the
test very flaky. Providing a fake "emulator" will make the behavior of
the test more consistent. As a bonus, using a fake ADB also speeds up
this test as the test now longer needs to wait for an Android system to
boot.

Bug: 77804802, 72748687
Test: CreateAndRunInstantAppTest
Change-Id: Iff0caa5cb562436f21af8e52f5517315a9c2aba4