Updating trebuchet
The latest trebuchet has a few changes most we don't care about. The
one we do care about is the long line fix. However the fix that is
currently in place throws away the line and this is invalid behavior

The follow up CL has the studio side fix for this so we don't diverge
from trebuchet.

Test: In trebuchet project
Bug: 77846431

commit c23425938328b3985034456ca9fe438120ae6adb
Author: John Reck <jreck@google.com>
Date:   Mon Apr 2 16:44:36 2018 -0700

    Fix line reader max length bug

commit 53938f3464e5009abd3d4a1403704b2a493dafca
Author: John Reck <jreck@google.com>
Date:   Wed Mar 28 16:58:26 2018 -0700

    Bump kotlin version

    Change-Id: I1f9e95e79bb153b6a16cf74c5717c5e447a0d622

commit 7028a231eb7337c430722b402b6e1c8ed53525b3
Author: John Reck <jreck@google.com>
Date:   Tue Mar 13 12:06:30 2018 -0700

    Update README

    Change-Id: I4459e8eec1d18df0785439cdc39ae926b2a7f792

commit 57a15330dd4b5a987b2a2c7803b3b2fdaf588aa1
Author: John Reck <jreck@google.com>
Date:   Tue Mar 13 12:05:28 2018 -0700

    prune sample_data

Change-Id: Iec09e9e2286ab86672978745e5e333820ad48564