[Gradle Build Model] Provide a method to copy a GradleDslSimpleExpressions
This CL adds two methods to GradleDslSimpleExpression and all its subclasses.
The first is getRawValue, this returns a value that can be passed into a
GradlePropertyModel to correctly set the value.
The second is copy, this creates a new copy of the GradleDslSimpleExpression
with the same value but no PsiElement. This allows the element to be reused
within the element tree without any problems stemming from the ordering of
PsiElement operations.

Test: existing
Bug: N/A
Change-Id: I4418df03940c987ad787b2a7d726593c33a34701