Fixing bug where delayed stopped/destroyed activities still render.
Sometimes when an activity was stopped the stopped event would happen
far later. This would result in the events timeline showing the
stopped then destoryed activity at the same time as the currently
running activity.

Removing this optimization fixes the issue by providing studio
with all data about activities and their states.

This change also has the side effect of fixing the bug when all the
way zoomed in events sometimes disappear. The symptom is different
but the cause was the same.

Test: Updated EventDataPollerTest
Change-Id: Iad8fab32344949480554c1688560f9e66f0e9a33
Fixes: 75984856