Fix race condition in IdeTestApplication
The IdeTestApplication constructor reflectively calls MainImpl.start,
which does its work on the UI thread via invokeLater. Recreating the
config directory after that has begun means it's possible that when
StartupUtil checks for the config directory's existence to decide
whether to show the config import dialog, it doesn't exist even though
it should.

Switch the order of the calls so that the directory is recreated before
the IDE is launched.

Bug: 77905470
Change-Id: Iec4aa20103408e87809cd667e0c3f7b91b3c3532