Add properties for the inflated view.
This CL allow us to show properties for a MaterialButton when a
Button is inflated as a MaterialButton in layoutlib.

The changes are made for both the property inspector v1 and v2.
Temporary: Use "src" instead of "srcCompat" since the
SimpleApplication is using Activity instead of AppCompatActivity.

Bug: 74326901
Test: Updated NlPropertiesTest to setup a view for AppCompat.
      Which is the same mechanism that MaterialButton would use.
Test-Strategy: ui.test.groups=THEME
Change-Id: I715ee61e7f3f8ee950221dc264ca5df55f1835a0