Don't wait for Bazel console to be active before clearing it
NlEditorTest#designEditorUnavailableIf* have been failing due to a
timeout where the tests are waiting for the Bazel console to be active
after opening it. These failures are caused by the warning that tells
you that BUILD files have been modified, which used to be a banner at
the top of the editor, but is now a notification at the bottom right
of the IDE that prevents the Bazel console from being active.

This CL resolves the issue by updating BazelConsoleToolWindowFixture
to no longer wait for the Bazel console to be active after ensuring it.
We don't need this wait, since we're already waiting for the console
to become visible and then are immediately right-clicking it. The
fixture now checks to see if the BazelConsole is visible and, if it
isn't, opens it via the "Tool Windows" menu rather than using
ToolWindowFixture#activate.

This CL also re-enables designEditorUnavailableIfLastBazelSyncFailed,
but leaves the other test disabled because it is flaking for an
unrelated reason.

Fixes: 77955344
Test: Ran designEditorUnavailableIfLastBazelSyncFailed 25 times
      locally with no failures.

Change-Id: I65ac0b039e4bdfd114d1ffa44baf3e5389f825a0