Determine the M1/M2 components based on the dependencies.
This reverses the previous approach where this was determined from
the app theme. However it seemed wrong that M1 components were
being presented in an app that already was depending on M2.
Now we determine which material components that are shown on the
palette by checking:
- The compile SDK, since M2 is available only for API 28+
- The presence of the design and material libraries

Bug: 74326901
Test: Updated and added unhit tests.
Test-Strategy: ui.test.groups=DEFAULT
Change-Id: Ic8cdc94dbc79e145e492eb1569636ce4bd112fad