Add getDeclaredDependency to AndroidModuleSystem
Project system's current dependency management APIs were designed
assuming a well defined closure of dependencies that will satisfy
all the needs of the IDE.  Because of this, dependencies are defined
as enums and the API doesn't allow creation of custom dependencies.
However as the list of needed dependencies grow, we can't always
predict what dependencies we need.

This CL is the first of a set of CLs that reimplement project system's
dependency management API to work with custom dependencies. Users can
now use maven coordinates to specify custom dependencies and still
remain build system independent.

This CL adds getDeclaredDependency which is analogous to
getDeclaredVersion. DependencyManagement.kt has been changed to use this
new method.

The old API is kept to preserve existing implementations until this CL
goes in.  A future CL will migrate existing usages to the new API and
remove the old API.

Bug: 77874749
Test: Created GradleModuleSystemIntegrationTest.
Change-Id: I59933d4f8426335cb9e472a4d745e51cbadd432c