[Gradle Build Model] Expand SingleArgumentMethodTransform
This change allows the SingleArgumentMethoTransform to be constructed
with multiple method names to recognize with the first being the one
that is used if the property doesn't exist yet or has been written to.

With will be used for File constructors to allow it to detect both
"new File" and "file" methods of construction.

Test: unit tests added
Bug: N/A
Change-Id: I5f40e07dda00786bfbd9b6754f6825e4c487c0a0