Tell users to sync when layout editor can't render
Currently, if the user tries to open the design editor or a layout
preview after a failed sync, the error message they get tells them
the feature is unavailable until a successful build, when in actuality
we require a successful project sync + source generation before
rendering.

This CL updates the error messages to avoid confusion. Requiring a
sync is likely a temporary solution while we figure out what exactly
needs to happen in terms of the build system in order to make
rendering possible (right now, sync + source generation is sort of a
catch-all). Ideally, we would be able to render something
no matter what and just present the user with a warning that some
views may be out of date, rather than completely blocking them from using
the layout editor.

Bug: 76490028
Test: N/A
Change-Id: If63556eea6d0c64f9efd1448e840acae38285a9d