App package name should be obtained from simpleperf trace
We were passing the application package name to SimpleperfParser
constructor, but the trace actually provide this data. Getting it from
the trace fixes the issue of package name not being available for
imported traces.

Also, we can infer the main thread name from the trace and pass it to
CpuCapture when we parse.

Bug: 74525724
Bug: 78222388
Test: updated SimpleperfTraceParser#mainProcessShouldBePresent
Change-Id: I6d66b9d07554a22e71eaab0cb763fd89742ea576