Updating System trace parser to throw exception instead of assert.
The try parse fail method of loading captures doesn't fail properly
when it attempts to load a systrace file. There are two reasons for this
1) We pass 1 for process id, sometimes the parser finds a process with
this id.
2) We assert but expect an exception.

This change passes AtraceParser.INVALID_PROCESS instead of 1 to
ensure we never get a process. As well as throw an exception instead
of assert to ensure the caller will catch us properly.

Test: AtraceParser updated with expected exception, also updated
CpuCaptureParserTest the previous test worked by a fluke. This test
now actually works.
Bug: 78250770
Change-Id: Idb90b695fcc3a727c6c4979d9a8add1605371ba9