Don't wait for sync from unsupported build systems
Currently, users with projects that aren't built with Gradle,
Bazel, or Blaze can't use the layout editor because it is stuck
waiting forever on a project sync. This is because rendering the
layout is blocked on ClearResourceCacheAfterFirstBuild, which
assumes there will always be an initial project sync.

This CL resolves the issue by having
ClearResourceCacheAfterFirstBuild skip waiting for a successful
project sync if there's no sync in progress and the sync status
of the project is clean. Since projects with unsupported build
systems are associated with DefaultProjectSystem (which always
reports a clean sync status), the layout editor will always
try to render for such projects.

Fixes: 78226465
Test-Strategy: ui.test.groups=DEFAULT
Test: Manually verified the Layout Editor renders for a Buck project
      + Manually verified behavior is unchanged for Gradle project
      + Bazel projects covered by NlEditorTest

Change-Id: I69d2bd1a1aafa221410f921ca6e2bc9c1610a9e4