Don't validate signed status for bundle output
We don't have a way to validate that the output of the bundle tool
is signed, so do not validate in the Run Configuration, as validating
prevents run/debug/deploy even if the output is signed.

In case the output is not signed, adb pm install gives an error
message. It would be best to have a dedicated error from Studio,
but this is currently not possible.

I opened bug 78602292 to track adding this feature to the Gradle
Model and Run Configuration.

Bug: 78567151
Test: TBD

Change-Id: I45237982407c08039667bfa68e4cc9455f97a900