Stop client if any involved thread terminates
If the client send or receive thread dies but the keep-alive thread
does not, particularly if this occurs after a test has started, the
server may hang indefinitely, since it uses absence of incoming messages
to detect client death.

This change causes the client IDE to close if any of the threads
terminates. The server will then be able to detect the client's absence
and start a new one.

Also catch Throwable instead of Exception in ClientReceiveThread to
improve diagnostics.

Bug: 78652452
Change-Id: Iccab74869e7c7a5ec38cf905aeab5c01de7eec5d
Test-Strategy: ui.test.groups=TEST_FRAMEWORK