Fix activate/deactive model
On the initial call to activate the model we were using the wrong source
so there would always be an activation for every model until the tab was
closed.
This CL tried to fix that and to make sure that we keep the models
active only when they are visible.

Bug: 78254948
Test: Existing tests pass
Change-Id: I1656ef08185771533d83bb091194b43379d6acd5