Replace FixedColumnTable
Because of the way FixedColumnTable inherited from JTable clients had to be
aware that column methods such as JTable.getColumnCount only took the
scrollable columns into account. Ideally the distinction between frozen and
scrollable columns should be invisible to clients.

Replace it with FrozenColumnTable which uses a composition model instead

Bug: Not applicable
Test: Existing tests
Change-Id: Id3a55465178abcf8436e0cf0945bfa630b39e0e7