Call sync once events from previous sync have finished onBuildCompletion
When a project has sync issues that cause a warning, we can display quck
fixes and some of them cause sync to be called again. If the user hits
the quickfix while current sync (or build) is still running, we record
that information by using setSyncRequestedDuringBuild and related
methods to call sync once the build finishes. Currently we call gradle
sync directly from onBuildCompletion, but if the build is not over yet
sync will not start.

This CL makes sure that there are not pending events from previous build
by calling gradle sync using Application#invokeLater.

Test: Existing test cover sync is succesful and manualy confirmed that
sync is enqued and then executed. Having a reliable test can be
difficult since this is a timing issue.
Bug: 76017112

Change-Id: Idff4268e65bbe46affba7c3ae3377da3b11111d9