Replacing the record icon with text "Record".
Also adjusted the paddings on the toolbar to make it easier in the
future to figure out what is going on. It was slightly offset due to
having a padding on the layout and in some cases a padding on the
elements. So I removed the layout padding and let each control manage
its 30px in the toolbar directly. This is in place of assuming 29.

Test: UI Change only
Change-Id: I3020d95e6c85c0daadc881952bf3fb2cad7963ff
Fixes: 77592513