Synchronizes per-device perfd thread execution
This change should fix the issue where the grpc InProcessServer
complains that the "name is already registered". The main problem
arises from not disconnecting the server when the device goes from
online -> offline (but not disconnected) -> online state.

After the fix, a new mechanism is needed to synchronize perfd
state modification. One way is to put all perfd modifications in
one giant thread, which would make connecting multiple devices on
start up slow. This CL implements a custom executor that
synchronizes state modifications on a per-device level. Also, to
save thread resources, it is able to safely clean itself up.

There are also some minor fixes to replace explicit synchronize
blocks with an implicit synchronized hashmap.

Fixes: 78526419
Change-Id: I11fd3912592320d38e7741b54ac613ecb7d0b17c
Test: Added test for the new executor.