[AVD Mgr] Fix bad UI scrolling when creating an AVD
When selecting hardware or a system image for a new AVD, the lists of
hardware or system images jumped spontaneously to the start of the
list.

This CL prevents the property change listener from scrolling the list.
With this change the list still scrolls correctly when clicking, tabbing,
and using the arrow keys.

Fixes: 78795961
Test: None added. ConfigureAvdOptionsStepTest still passes.
Change-Id: I996cd42ff3f99a34fb04d0b0f1398f1ca54f29b6