Add UI test to make sure file can be open without highlight errors
This is a test to avoid regression for R class highligh error. Due to
class generated is invoke by blaze sync which we have not yet find a
good way to do integration test, I create this UI test instead. So we
will check MainActivity.java which suppose to be open without any
highlight errors. This test may not only catch R class related error,
but this test result should be good for identify R class regression.

Change-Id: I88957cc201c47e45d1775a78269355c72d9ab292
Fixes: 79371008
Test: BazelStartupSyncTest