Walking components and setting default background color for tooltips.
Simple change to tooltip view to set the default background color on
all tooltips. This removes the need for each component to set its own
background color.
In the future we will also want to do this with font.

Also adding the unavailable details to the default content view so
it properly gets the background color set.

Test: None
Bug: None
Change-Id: I81d45e0023ec94527043f12fe84b3ef7b531db42