Fix installed app launch when instant app library present in SDK
Unfortunately this logic was meant to exclude instant apps when that
library is present, but ended up excluding all apps. Luckily the library
in question isn't in the SDK yet. See bug for more context.

Follow-up question: Is it possible to disable a StudioFlag for older
canary versions, so they won't break when we do roll out this
integration?

Change-Id: I2dd32879761e333cb84c779037c1941f72ff81cd
Fixes: 79491722
Test: Verified manually. There seem to be no tests at all for AndroidLaunchTasksProvider.