Fix selection "creation event" not firing a second time
Currently, if you create a selection range by dragging it, and then
immediately create another range by dragging in a new area, we only
send the "selection created" event for the first selection, not for
any subsequent ones.

This was due to the underlying model not able to tell that the
parent component was creating a new selection, not simply
modifying an existing selection. Now, we clear the selection
model to indicate our intentions explicitly.

Bug: 74004663
Test: Added
Change-Id: I911794ee3722f4cf6d71ff5eac4f49d48fba97a4