[AVD Mgr] Don't crash when launching a skinless AVD
When we launch an AVD that shows a device frame (a device skin),
we take that opportunity to verify that the latest skin version
is being used.

When we launched an AVD that does not use a skin, our check of
the skin version NPEd for lack of a skin path.

Don't check the skin version if there is no skin path.

Fixes: 77543709
Test: Added AvdManagerConnectionTest.testStartAvdSkinned() and testStartAvdSkinless()
Change-Id: I917d45e924cfed0ccbe8ef3e7f752bbee9d7b528