Importing a systrace prompts the use to select a process on parse.
Everytime the trace is parsed we will display a dialog to the user
prompting them to select the process they want to analyze via systrace.
This only happens when a user imports a systrace file. What I mean by
every time is, each time the user leave the capture then reselects the
capture we again ask the user what process they would like to select.
In the future we will save the users selection.

To generate the list of processes the user has to select from we first
parse the capture into trebuchet models then sort the models based on
a few heuristics.
1) hint, in this case the package name is used as a hint
2) # of threads.
3) do we have a valid process name
4) everything else.

Test: Updated the logic in AtraceParser, CpuCaptureParserTest
Test: TODO on testing dialog window, not sure how.
Bug: 79578611
Change-Id: I43fa9f626acf37910fb5a3c02a04fc008a3feef7