[SVS] Switch build variants from UI - CL1.
When user changes build variants from UI, there are three
different cases,
1. Build files was changed since last Gradle Sync.
2. Build files are not changed, the selected variant was not
requested before.
3. Build files are not changed, the selected variant has been
requested before.

This CL handles case 1 and case 3, case 2 will be in next CL.

The workflow is,
1. Starting from the selected module, find all dependency modules
that needs to be updated.
2. Update the selected variants in AndroidFacet and AndroidModuleModel
for all modules in step 1.
3. If build files have been changed(case 1), request a full gradle
sync. The Gradle Sync should be using the newly selected variants
as that have been updated in step 2.
4. If build files not changed, and selected variant exists(case 3),
do module setup for all modules in step 1.

Bug: 79384659
Test: BuildVariantViewTest and BuildVariantUpdaterTest.
Change-Id: Ic163f742bfbb44027c930fc5b070613da70a95ba