Increase threshold for image diffs to enable android-tests on Foundry.
This CL fixes the last remaining test failure on android-tests by increasing the
threshold for image diffs in UI tests from 0.5% to 1.0% wherever we have seen
failures on foundry. The images are different because the foundry Java toolchain
uses a different JDK then the one we use locall, i.e. the one that is in
prebuilts/studio/jdk.

We need to fix that and I've filed a bug here b/79948364. In the meantime, it's
important we move forward our migration to bazel. This is the largest
no_test_foundry target (in terms of run time). This CL also puts the image diffs
in the undeclared output directory so that they are archived by ResultStore.
They can be found in the Artifacts tab, under the specific failing target.

Bug: 72180168, 73948555
Test: ./tools/base/bazel/bazel test --config=remote
           //tools/adt/idea/android:intellij.android.core.tests_tests
Change-Id: I6f40d4f2f4aa20120e22e83d8266fdf3673f5938