Support drag+drop and double-click to open hprof/trace files
In addition to the already supported "File -> Open" flow, we also want to
be able to open .trace and .hprof files using the Project Structure panel
in Android Studio. We can either double-click this files or drag-and-drop
them, and that will open the Android Profiler window and import the target
file into it.

Drag-and-drop also works when dragging the files from system's file viewer
(e.g. Finder on mac) into the IDE.

Bug: 78888720
Test: Manual. Open an hprof/trace file using File -> Open, double-clicking
      it in the project structure panel or dragging them into the IDE. That
      should open an Android Profiler panel and import the trace file into
      it.

Change-Id: I78b030b5f4a5a167c8bde264fbee3a73ef16648a