Allow Layout Editor to render after Blaze partial success
Large first-party apps are often in a state where most of the blaze
project syncs successfully, but a handful of failing targets keep
the project sync from being considered successful. These "partial
success" cases are considered failures by the project system sync
APIs. Since this situation is really common and the layout editor
requires a successful sync to render, the end result is that most
of the time first-party app developers can't use the layout editor.

This CL changes the project system APIs to consider "partially
successful" Blaze syncs to be successful. This effectively removes
the successful sync requirement for layout editor rendering, as the
Blaze plugin reports any sync failure due to compilation or BUILD
file errors as a partial success. This may lead to issues when using
the layout editor when trying to access a component for which sources
have not yet been generated.

This CL also disables NlEditorTest#designEditorUnavailableIf* until we
figure out a better way to handle partially successful syncs, as both
of these UI tests expect the layout editor not to render after a
partial success.

Bug: 78228400
Test: Manually confirmed that layout editor renders for memegen
      after a partially successful sync.

Change-Id: I7a934884a6f10e2f5f131ac68a2a4a1faebf0103