Specify allowed dependencies during test init.
Currently tests that use TestProjectSystem are allowed to add any
dependency they want to add, regardless of their actual existence.

This CL mandates that all dependencies to be used in a test should be
pre-declared to the TestProjectSystem upon initialization.  With this
new approach, testers will need to understand what dependencies the
tested code requires.

E.g. If the code being tested adds appcompat to the project during test
execution, then appcompat should be declared to TestProjectSystem during
init. This means if a piece of code tries to add a dependency it did not
require before, a test will be able to catch it.

Bug: 80130524
Test: All affected tests have been updated.
Change-Id: Icfc7ea90eb337626f6eb1e4324c8d3394120665f