Add ability to change listitem multiple times
This CL also adds the changes needed to persist changes to the
itemCount.
If the use selects a new template for the RecyclerView and DOES NOT make
changes to the template contents, the RecyclerViewAssistant will
identify the template and pre-select it in the list of templates. This
allows for a more consistent behaviour with the rest of the assistants.

Bug: 79115580
Test: Updated UI test
Change-Id: Ib6c810b5d9aa8a4519a643cf9ba53cb6b06783a2