Fix refactoring for overlapping kotling packages
When refactoring to AndroidX Kotlin code, the refactoring tool will find
the references to the last segment of the package instead of the full
package name.
For example, for "import android.support.v4.media.Test", the package
refactoring will find "media" as package instead of
"android.support.v4.media".
This is just the way the PSI for Kotlin works. The tools relies on text
ranges to find when something has been already been refactored. Because
of the wrong use of ranges in Kotlin, the refactoring fails to recognize
those overlappings.

Fixes: 80188241
Test: TBD, probably as a UI test
Change-Id: I1750647ff320bbf79547c4a208d36841a2ea92f2