AndroidX refactoring: recognize class names in attribute values
- Added ResAutoAttributeValue and extended
AndroidResourceDomFileDescription to add a way to work with res-auto
attributes.
- Added app:actionProviderClass and app:actionViewClass to the DOM model
for MenuItem by hand, so the converter can be specified. For now they
don't specify the allowed superclass, because of a bug where package
private implementations from the support lib are found. This will be
fixed in a follow-up CL.
- Started handling references from XmlAttributeValue in
MigrateToAndroidxProcessor.

Test: MigrateToAndroidxTest
Bug: 80153626
Change-Id: I2c4b5c3977fddf20eb2b2e6be70b3e56db4b6db4