Stop returning MessageBusConnections from project system sync APIs
ProjectSystemSyncUtil#listenForNextSyncResult and
SyncUtil#listenUntilNextSuccessfulSync both return instances of
MessageBusConnection to allow callers to stop listening for sync
results at any time by disposing these connections. This is an issue
because the APIs also dispose of the message bus connections under
certain conditions. This makes it very easy for callers to end up
in a scenario where they are holding a reference to a disposed
MessageBusConnection, rather than letting GC reclaim the memory.

This CL resolves the issue by inlining the MessageBusConnection
disposal logic in the calling classes that make use of this
functionality. Since there are no other usages of the
listenForNextSyncResult API, it has been deleted. The
listenUntilNextSuccessfulSync API has been updated to return Unit.

Fixes: 80083771
Test: updates to ClearResourceCacheAfterFirstBuildTest
     + SyncUtilTest and GradleProjectSystemSyncManagerTest still pass
     + LayoutEditorMemoryUseTest (cherry picked from If1376290f)

Change-Id: I7b7d9523cca60aa2e1527480ee97d408d737c8d2