Showing a warning when systrace is missing parent_ts field.
Pretty much the only way for us to know if data is missing from systrace
is to check to see if the timestamp markers written by atrace when
start is called are in the buffer. If these fields are missing
then we know the capture hit its internal buffer size and started to
write over itself.

See b/78191316 for more information.

Test: CpuProfilerStageViewTest checks for dialog, AtraceParserTest
validates isMissing
Fixes: 78191316

Change-Id: Ia80800a50709c4e024b3e0ba330f1c59085c4540