Fix Record button sometimes being enabled in dead sessions
We were correctly disabling the record button when the CPU stage was
created in dead sessions. However, we also change the enable status
of the button when transitioning between states (e.g. going from PARSING
to IDLE).

Navigating to a capture using the sessions panel means we're going to
change states after initiating the stage, as the capture still needs to
be parsed. As a result, the record button was being enabled in such
scenario and clicking on it was throwing a NullPointerException.

Instead of blindly setting the enable status of the button to true, we
need to check if the session is alive first.

Bug: 80418632
Test: Manual. Following the bug description, create a session and record
      a capture on it. Then, start a new session and navigate to the
      capture recorded using the sessions panel. The record button should
      be disabled.

Change-Id: I9cb458df9d26b6da05409bed8015c33c7222744e