Ensure transmitted Throwables are Serializable in JUnitClient
Attempting to send a non-serializable TransportMessage will result in
client death. The only thing I've seen be a problem is the Throwable in
a Failure object.

After this change, the client will attempt to send the message, and if
it fails due to a serialization problem, will retry with the exception
replaced by a plain Exception whose message is a string representation
of the original.

Bug: 80371267
Test-Strategy: ui.test.groups=TEST_FRAMEWORK
Change-Id: I150e2af3187ca973389ce0507c2ab75030329539