Create uitest_deps maven repo
When running UI tests on the release, test project dependencies must be
bundled in the gradle/m2repository, since tests should not depend on
remote servers.

Most of the dependencies are covered by the already existing maven repo
//tools/adt/idea/android:test_deps. Additional dependencies will go in
//tools/adt/idea/uitest-framework:uitest_deps.

This change creates this new maven repo; more stuff may need to go in
there but this is a start. It also adds bazel invocations to build these
repositories to the bazel-dependencies ant target.

Another change in this topic changes where the gradle zip is placed to
match that expected by
EmbeddedDistributionPaths#findEmbeddedGradleDistributionFile, which is
now used in GuiTestRule.

Replace a dependency on an old version of mockito in a test project with
a '+' version.

Bug: 109805478
Bug: 109812409
Bug: 74197807
Test-Strategy: ui.test.groups=TEST_FRAMEWORK
Change-Id: I45f8089f599f65c5c7786c6875d5fb85c721dfcd