Add a new getResValues() method to AndroidModel
Move the gradle-specific logic in
DynamicResourceValueRepository.getFullTable to a new getResValues()
method on AndroidModel, as a first step to providing this functionality
in a build-system-independent manner.

This restores the code reverted by commit
3688df07c8aadd69e7b6dd00796b5402a1aeb8a7, originaly implemented in
commit 5f5059a37775876c67afd6d5b3d94eb7694f179b. However this patch
additionally includes a fix for a bug introduced in DynamicResourceValueRepository.

The original patch was applying the resource overloads in an incorrect
order, causing the flavor values to override the build type values.
That caused a test failure NlPreviewTest.testRenderingDynamicDependencies.
This patch reverses the order and fixes the failing test.

Bug: 109665898
Test: NlPreviewTest
Change-Id: Ibe73f2bc38babe564ad939e2ced75a4fddbb58d3