Rename profiler spec images
This change will be invisible to the reader of the doc, but we rename
the images to be more descriptive, instead of simplying numbering them

e.g.
  memory-alloc-managed.png, memory-alloc-native.png
instead of
  memory1.png, memory2.png

The new naming convention will make it easier to maintain this
document moving forward, especially if sections ever get reordered,
but at least new people adding images can just name them directly,
instead of searching around for the next available numeric suffix.

Bug: N/A
Test: N/A
Change-Id: I24c0419ffe882b8feba91feb2f7d14153ee213e6