Rename CPU profiling configurations default names
We've decided to change the default names of CPU profiling configurations.
The new names differentiate ART sampled and instrumented captures, and that
introduces a technical challenge because the trace info didn't have the
information regarding the mode (i.e. sampling or instrumentation) used to
generate the trace.

We had to refactor our APIs and some of our services in order to introduce
the "mode" information to traces info and CpuCaptureSessionArtifact can now
use CpuProfilerMode, in addition to CpuProfilerType, to determine the name
of the artifact displayed in the sessions panel.

Bug: 77912241
Bug: 110105335
Test: updated existing tests and add some more
Change-Id: I05cf361e41c1f35e19c07eca452aa16b498160cb