Refactor error validations.
Move the validate method to a more general EditingSupport.

Also change the method to a lambda, the idea is to provide lambdas
for specific data types i.e. a Dimension would require a number
followed by optional units (px, dp, sp).

I intend to expand the EditingSupport to include support for
completions in a future CL.

Bug: 73778668
Test: Exisitng tests should pass.
Change-Id: Idb3395cc3ae850f35cbfa5cce3a1639c293573e2