Implement fly-away in table.
The "Declared Attributes" table is showing the attributes that
have an explicit value in the XML file. We want the items in
this table to update whenever:
- a value is set on an attribute that previously was empty
- a value is removed from an attribute that previously was set
- a value is entered for a new attribute in the table

In the last case the attribute row should "fly-away" into its
proper place in the sorted table.

Note that the first case only can happen when the attribute
value is set in the "All Attributes" table.

Bug: 73778668
Test: Unit tests added and updated.
Change-Id: Ib26b5240469b11991b6177f9b1c85e9bb2e8549a