Caching the last run app to improve auto-profiling behavior
Previously, if the user opens the profiler after they run an app,
the profiler will not know about the app and hence auto-profiling
won't kick in. This change caches the info related to the app that
was last run, so profilers can try to start profiling right away
if the user opens the window manually.

Bug: 110233760
Test: N/A

Change-Id: I1ec257ecfd391f8d5e2c2382ecd8676aa64cda99