[SVS] Add a flag to enable/disable single-variant sync.
Add the flag to Gradle experimental features form. This
setting will be per-project.
Gradle sync should switch to new gradle sync and perform
single-variant sync if this flag is enabled.
This flag is made invisible for now to avoid accidental
release to user. Should be made visible when the feature
is complete.

Keep the flag in StudioFlags for now, which enables single
variant sync globally, and is currently being used in tests
and also for debug purpose. Eventually the one in StudioFlags
should be removed.

Bug: 110579535
Test: ExperimentalSettingsConfigurableTest
Change-Id: I4ca9db1a8265ff78a65cd9e00509bf753d5f468c