Catch exceptions when trying to read passwords for signing
When the user tries to build a signed APK, the keys are read from a
keystore. The passwords for this store can be saved to reduce user
interaction. In the case of an exception while retrieven the password
from the PasswordSafe, signing cannot continue and causes an error,
leaving the user without this option to sign the APK.

This CL catches the exceptions generated by the password safe and allows
the user to continue signing after introducing passwords manually.

Bug: 70654787
Test: Manually caused an exception while instancing the password safe
and confirmed that signing is still possible.

Change-Id: Id90f0402965beab7e2be63762bded1ddb6c22365