Optimize remote SDK repo reloading.
Remote repo reloading implies establishing an actual TCP connection
and downloading a number of XMLs from repo sources, which is quite
an expensive operation to be done on every UI update. So this CL
optimizes out some of the cases when a plain local repo refresh
can be used instead (e.g., on pressing 'Apply' unless any remote
sources have changed), or can be avoided altogether (e.g., when
no relevant changes have been made to the sources tab but it needs
a plain UI refresh).

Bug: b/110928872
Test: To be followed up. Producing a test class for these panels
needs to be done from scratch.
Change-Id: I933b87eaf473afee60039c137e41f003a844e263