Fixes for renaming resources and light R classes
When light fields are used, they fail PsiManager.isInProject check,
which makes PsiElementRenameHandler.renameabilityStatus consider them
impossible to rename. This makes sense, since conceptually the field
itself should not be renamable, it should trigger renaming of the
underlying resource. For now our rename handler and processor have
dedicated code for handling fields though, so to not change everything
at once we wrap and unwrap the light fields in a fake, writeable PSI
element that gets passed to the processor.

This code needs serious changes, as currently you get different results
depending on whether you started in a java file or XML. I didn't want to
block light R classes work on this, since this is a big chunk of work of
its own. So for now we just cheat a bit with the wrapper.

Bug: 77801019
Test: AndroidRenameTest
Change-Id: Iec112582b15f7c009076ae49dd160daccee8b5b7