Decrease scope of the workaround in CpuCaptureViewTest
There is a workaround in CpuCaptureViewTest and it is in the common
method of all tests, however it applies only for ATrace related tests.
So, by decoupling the Atrace related test into its own test, we can move
the workaround as well. By doing so, we decrease scope of the
workaround.

Test: refactoring of a test
Bug: N/A
Change-Id: I9c209b08b97926625ae19aa22237d3997d399de6