Build proto AAR resource repositories based on the model
This way we get a somewhat working project now that the build system
can rewrite AARv1 files to be namespaced.

Existing logic for finding AARs in AppResourceRepository has been
replaced by per-project-system code that we've written for light R
classes. This way we have enough information to create a repository of
the right type.

Note that the model is correct but the res.apk files don't exist until
the first build, since the Gradle generateSources task is not set up
correctly yet.

Bug: 74425399
Test: existing, AppResourceRepositoryTest
Change-Id: Ic52bb6b82b2613684f2682356dfd4efde98f6dc4