Introduce FakeCpuProfiler
We have a lot of logic in tests that does the same thing
again and again. Also, we don't have convenient methods which we can
use to simulate the common events in CPU profiling, e.g capturing a trace.

This CL introduces a class that can be used by CPU profiler tests
to simulate CPU events.

Test: refactored the existing test to use it
Bug: 111172562
Change-Id: I4b13187efc03c0aa9e6fed516b52566b59bb8867