Don't cache namespaces for unknown VirtualFiles
The underlying index layer caches most of the computations, so getting
the list of corresponding order entries should not do any IO. As far as
I can tell, computing the namespace is only done when the user tries to
navigate to the resource definition, we don't continuously run any
visitors on non-project files. Also this saves from bugs when the cache
is not invalidated correctly.

Bug: 109827637
Test: existing
Change-Id: I0a5c182a95e811fc4b5ee330f698f4c4094a77c4