Don't throw exception when parsing unexpected method names
We were throwing exceptions in NodeNameParser everytime we found an
unexpected method name, more specifically without matching parenthesis
or angle brackets.

As we parse the symbols by hand, we might miss some cases and mess up with
the symbol name and its angle brackets/parenthesis. As a result, we would
crash and abort parsing the capture because of that. Instead, we should
fallback to the original name if we can't manipulate it properly.

Bug: 77945714
Test: NodeNameParserTest#unmatchingAngleBracketsShouldntThrowException and
      NodeNameParserTest#unmatchingParenthesisShouldntThrowException

Change-Id: I21306f6f110b7008bcb67c995781cb4a7a39f715