Fix attr completion in namespaced projects
StyleItemNameConverter now highlights unknown attributes, even outside
of the "android:" namespace.

Also converted StyleItemConverter to Kotlin, as I had to rewrite most of
it anyway and handling nulls was easier this way.

I renamed the test, as it was referring to the wrong non-namespaced
class (I got confused by the names earlier).

Bug: 111507468
Test: AndroidNamespacedValueResourcesDomTest.kt
Change-Id: I840e8787017cdea1553761adc53709be17c1773e