Allow preview deps from layout editor palette.
Currently clicking on the "download" button in the Palette of the layout
editor fails when:
1. The project uses androidx dependencies.
2. The palette item's androidx dependency is in preview.

The failure is caused by palette2.DependencyManager trying to add
a artifact that doesn't have a stable dependency while setting the
"includePreviews" flag to false.

This CL fixes this situation by setting that "includePreviews" flag to
true when palette2.DependencyManager tries to add an androidx
dependency. The logic for adding non-androidx dependencies aren't
changed.

Summary of behaviours after this CL:

* Adding recyclerview in a project already using androidx will add
a preview version of recyclerview (there isn't a stable version).

* Adding recyclerview in a project not using androidx will add a version
compatible with the test of the dependencies like it did before.

* Adding recyclerview in a project with no dependencies will depend on
the status of StudioFlags.NELE_USE_ANDROIDX_DEFAULT. This is also the
default behaviour.

Bug: 111124139
Test: Verified that dependencies are added correctly according to the
version that is currently in the project. Verified with older projects
that palette behaviour doesn't change for projects with non-androidx
dependencies.  Unit tests to be added in subsequent CL.

Change-Id: Id3fd79afc43ba57e2934368730f265bca5dae397