Fallback to preview when stable deps aren't available
Currently clicking on the "download" button in the Palette of the layout
editor fails when:
1. The project uses androidx dependencies.
2. The palette item's androidx dependency is in preview.

The failure is caused by palette2.DependencyManager trying to add
a artifact that doesn't have a stable dependency while setting the
"includePreviews" flag on DependencyManagementUtil.addDependencies to
false.

This CL fixes this situation by changing includePreviews to
preferPreviews. Setting preferPreviews to false will first search for
a stable version of the dependency and in the event that a stable
version does not exist, addDependencies will fallback to searching for
a preview version of the dependency and add that instead if it exists.

Summary of behaviours after this CL:

* Adding recyclerview in a project already using androidx will add
a preview version of recyclerview (there isn't a stable version).

* Adding recyclerview in a project not using androidx will add a version
compatible with the test of the dependencies like it did before.

The CL also changes NavDesignSurface's previous includePreviews=true to
preferPreviews=false, as this preserves the correct behaviour.

At this time there are no more users of the preferPreviews flag and it
can be removed in a follow up CL.

Bug: 111124139
Test: Verified that dependencies are added correctly according to the
version that is currently in the project. Verified with older projects
that palette behaviour doesn't change for projects with non-androidx
dependencies.  Unit tests to be added in subsequent CL.

Change-Id: If90e5e1d18779031ac7e167686c23af7389eb2ff