Replace layout inspector debug actions with a test fake
The layout inspector works by connecting a debugger to an Android
application process. The debugger sends commands to the process to
retrieve information about the process, in particular layout
information from the process. This requirement causes the
LayoutInspectorTest to contain a dependency on the emulator.

The dependency can be removed by using test fakes. However, implementing
a fake only with FakeAdbServer would be impractical, as
the fake would need to re-implement a lot of functionality in an Android
system. For example, some of the actions the fake would need to cover
are:

* recognizing different debugger commands to respond with the right
response.
* handling debugger commands coming in asynchronously.
* parsing debugger command packets to retrieve data from the packet.
This data is generated at runtime, and the debugger fake would also need
to build responses based on this data. The prime example of this case
are packet IDs that are generated within Android Studio.

Based on the given requirements, a debugger fake would also have to
parse recorded responses from a real system. The fake would need to
recognize which fields need to be swapped out with data that is
generated at runtime, such as packet IDs.

Additionally, a debugger fake may have to return different responses for
the two instances of the same command.

It might be possible to inject a mock somewhere in the project such
that the test can control the dynamic information that is generated from
the IDE, but it is easier to replace the objects that handle debugger
actions on a high-level (e.g. actions that don't need to worry about
JDWP, packet IDs, or how to talk to the debug server) rather than trying
to mock out low-level responses. The lower level functionality can be
covered with smaller integration tests than a complete UI test.

The above requirements make implementing a debugger fake a much more
daunting task. It is easier and more effective to use mock objects
to replace real objects within the IDE. This CL is a demonstration of
how to use mocks from a UI test to mock out the debugger dependency for
the layout inspector functionality.

Bug: 74079626, 79582420
Test: bazel test //tools/adt/idea/android-uitests:LayoutInspectorTest
Change-Id: Id0d8b978da61c8e186b6a106f2e93128477a715c