[UI Tests] FirstRunWizardCancelTest failing after latest IJ code merge
Running test from IDE throws an exception RestoreFlagRule,
replacing by an alternative tear-down.
Seems that the new merge as changed the amount of time it takes
for transferring the focus from a closing window to the next
visible one, so we request focus manually (this is a PSQ
window manager problem only).

Fixes: 111423426
Test: Run FirstRunWizardCancelTest on Mac and PSQ
Test-Strategy: ui.test.groups=PROJECT_WIZARD
Change-Id: I98de4950475831e3b4d1759b4054374739b2a54e