Add network caching support to SDK downloads.
This CL adds the ability to tune HTTP caching in StudioDownloader
depending on the calling context.

This change is primarily targeted to support users who work behind
a proxy with caching enabled - this should considerably speed up
their downloads. A typical example of such a setup is a corporate
environment, therefore potentially a large number of people who are
behind the proxy would benefit from such on-proxy caching.

I've verified that currently dl.google.com does not restrict caching,
and sets expiration time to a non-zero value (24 hours). At the same
time, client-side HttpRequests issued by the IDE do add no-cache headers,
thus insisting on disabling the caches unconditionally. This change
therefore makes the client-side configuration more flexible to
accommodate caching when it is available and appropriate from the
calling context perspective (e.g., SDK package download would be OK
with caching, whereas a check for update would insist on no caching,
because the manifests could be updated on the original server).

For the implementation logic and Downloader interface change see
the other CL in the topic - ag/4578323.

Test: Added StudioDownloaderTest.testHttpNoCacheHeaders, which
fires up a compact http server and provides facilities to verify
the request header contents.
Bug: b/111470001
Change-Id: Ie92ab3e6e9e1fbf97176e764b3410974e896199c