Add popular NPW project generator
This CL adds a tool which generates NPW projects from "popular" input
combinations. This will be used later to generate shipped Gradle Sync for
NPW projects.

Currently this is in testSrc/ because it depends on test infrastructure.
Actually this is not a test but a tool and it will be moved to src/
later.

Test: N/A
Bug: 80116542
Change-Id: I507d3d3c34bd029b14457749d9fc652ab26c7eca