IndexingSuspender no longer enters dumb mode
The platform has a lighter-weight mechanism for preventing indexing
during heavy file-changing operations (such as those performed when
checking out a branch, etc). DumbService now supports direct
suspension via BatchFileUpdateEvents, as of 2018.1:
https://github.com/JetBrains/intellij-community/commit/0c051e7ca41048e2bc2eec4a7a870c1fdd0525fc

We switch to use the same mechanism to prevent re-indexing:
  * further simplifying the custom logic in IndexingSuspender
  * avoiding an entire class of problems by not entering dumb mode
(we have dealt with quite a few bugs related to symbol resolution,
finding test classes, etc. which do not occur in IntelliJ, but only
in our case because we might still be in dumb mode when background
tasks are started, such as code highlighting, etc).

Bug: 78547457
Bug: 69417268
Test: existing
Change-Id: Ic59763eeb7ea3f7f1a8637fbdcdb311f9530b7be