DefaultRecipeExecutor: Fix NullPointerException crash report
When adding project dependencies from the Firebase assistant,
the Template Recipe is run without any of gradle Template
values (eg ATTR_BUILD_API), that we were relying to check
if a project supports androidX.
This CL also fixes a NullPointer Exception when the project
gradle file is completely empty and we add a dependency to
it. This happens for Unit Test projects.

Fixes: 111832989
Test: Added TemplateRecipeTest
Change-Id: I02be1b787f0df22f9e4a332a16315f6871fe76a5