Introduce StatusPane to CpuCaptureView
The new CPU profiler recording flow includes a view in the CPU details
panel displaying the current status of the profiler (e.g. recording,
parsing) in order to increase discoverability.

In this first pass, we are introducing this panel for the recording state.
Later, we will expand the solution to include the parsing state as well.

Bug: 111779496
Test: CpuCaptureViewTest#whenRecordingThereShouldBeInstanceOfRecordingPane.
      Updated CpuProfilerStageViewTest#recordButtonDisabledInDeadSessions as
      well.

Change-Id: I87214d0b119f12062aa955dda3a7eb58fd773156