Simplify logic that determines visibility of ConnectDebuggerAction
Instead of !getFacets(Android).isEmpty(), use hasFacets(Android). This
removes the need to iterate through all the modules unnecessarily. As an
additional tweak, we also unconditionally show this in Android Studio.
In large projects (> 500 modules), this particular method was showing up
in freeze logs (e.g. 10% of freezes in b/112166795 were due to this),
mostly because this is invoked from a periodic timer.

Bug: 112166795
Test: manually confirmed that the action is enabled inside Studio.
Change-Id: I9f9a50104aa7041d66ec2756b18c130164d63803