Fixing up disposal in NavTestCase subclasses
We were unnecessarily using a custom root disposable. Switched to using
the project.
In some tests we were explicitly disposing things at the end rather than
registering with the disposer. That meant that if the test failed, the
project would be leaked which causes the test framework to report errors
for everything.

Test: This is a test-only change
Bug: none
Change-Id: I201e7dbc709f15d2f785bb4b0c38e422ecdf8fd1