Check for existing dependency in ConvertToConstraintLayoutAction
In order to convert a layout to ConstraintLayout, the module containing
the layout xml must depend on constraint layout. Currently,
ConvertToConstraintLayoutAction tries to ensure this by attempting to
add the dependency and failing immediately if the operation which adds
the dependency wasn't successful. But since adding dependencies
automatically isn't supported for Blaze projects, this feature will
never be available in ASWB, even if the dependency on constraint layout
is already present.

This CL resolves the issue by having ConvertToConstraintLayoutAction
check first to see if the module depends on constraint layout before
trying to add the dependency.

Test: manual (for both Gradle and Blaze projects)
Bug: 112323173
Change-Id: Iaa481a540a11887ef0fbc66062acbad181372026