Code restructuring
The initial code built pathData string and gradients separately, which
implied calculating the absolute position of the shape twice and making
the same kind of calculations at different stages of the code.

This change list comes with a more structured way of computing the
shapes and a more independent shape model.

Flow:
SketchParser reads the input file and returns a SketchFile object
A list of LightVirtualObjects is returned after calling generateFiles()
	Pages are retrieved from the SketchFile
		Artboards are retrieved from the SketchPage
			VectorDrawableFile is created
			DrawableShapes are retrieved from artboard
				ShapeModels are created
					pathData Strings are generated
					transformations are applied
					operations are applied
				ShapeModels are turned into DrawableShape
			DrawableShape is added to VectorDrawableFile as
			path

Bug: N/A
Test: VectorDrawableFileTest.java
Change-Id: I2eda54896083b01f47033099491defff23e9ae6a