[SVS] Support of ndk modules in variant-only sync.
Variant-only sync happens when the selection in Build Variant
window is changed, the build files are not changed, but selected
variant is not synced yet. This CL handles the change of variant
selection for native modules.

Since native modules also have AndroidProject model, we request
for NativeAndroidProject and NativeVariantAbi only when AndroidProject
is available. The dependency modules are defined by Variant
model, rather than NativeAndroidProject or NativeVariantAbi, follow
the logic of Variant model for dependency modules. The abi name
should be consistent with dependent module.

Add NdkVariantChangeModuleSetup to do module setup after variant
change for Ndk modules. Also, rename VariantOnlySyncModuleSetup
to AndroidVariantChangeModuleSetup to be consistent and understandable.
Add steps to call NdkVariantChangeModuleSetup in ModuleSetup
after sync.

Bug: 79581802
Test: VariantOnlyProjectModelsSetupTest, BuildVariantUpdaterTest.
Change-Id: I7f3d6ade3c08198d1e06962f67fc9af54541794d