Convert network overview tab to use a pretty-printed editor
The pretty view is a nicer default in a majority of cases,
but in a small subset of cases, users may need to know
EXACTLY the format of the payload sent to them by the
server, so we opted to show the raw response.

However, we've since introduced a toggle in the request and
response tabs which allow users to see a raw view of their
payload there. Therefore, we should make the overview tab
use the pretty view instead.

Bug: N/A
Test: N/A, visual change only.

Change-Id: I3c84e0b32f70b357b084a4111b3bbf4c81637060