Room: Fix completion in UPDATE queries
The generated parser was not smart enough to create an UpdateStatement
object for invalid queries starting with `UPDATE`, which left us with
a flat list of tokens and no information to offer code completion on
incomplete queries. The fix is to manually add the `pin` option to the
relevant rule. This is most likely because the `UPDATE` keyword is used
in other places (e.g. `INSERT OR UPDATE INTO foo ...`) in the grammar.

Change-Id: Ica3a5cfcbc3b9f51755cf6ab497b0c4cc56e776d
Fixes: 80198036
Test: RoomSqlParserTest.kt, ColumnReferencesTest.kt, TableReferencesTest.kt